requirements.txt
appdirs==1.4.4
black==21.7b0
bleach==4.0.0
build==0.6.0.post1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dataclasses==0.6
docutils==0.17.1
idna==3.2
importlib-metadata==4.6.4
keyring==23.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pep517==0.11.0
pkginfo==1.7.1
Pygments==2.10.0
pyparsing==2.4.7
readme-renderer==29.0
regex==2021.8.21
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
tomli==1.2.1
tqdm==4.62.2
twine==3.4.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.0

setuptools>=57.0.0
deploifai>=0.0.1
PyInquirer>=1.0.3
dvc>=2.7.4
PyYAML>=5.4.1